// src/view/main/signup-login/components/enums.ts
export enum CredentialFieldNames {
  Email = 'email',
  Password = 'password',
}

export enum PersonFieldNames {
  FirstName = 'firstName',
  DateOfBirth = 'dateOfBirth',
  LastName = 'lastName',
}

export enum AddressFieldNames {
  CountryCode = 'countryCode',
  City = 'city',
  Street = 'street',
  PostalCode = 'postalCode',
}